+++ src/main/java/net/grid/vampiresdelight/common/block/BlackMushroomBlock.java
@@ -8,11 +8,16 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.MushroomBlock;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
 import net.minecraftforge.common.PlantType;
 
 public class BlackMushroomBlock extends MushroomBlock {
+    protected static final VoxelShape SHAPE = Block.box(5.0D, 0.0D, 5.0D, 11.0D, 8.0D, 11.0D);
+
     public BlackMushroomBlock(Properties properties) {
         super(properties, VDConfiguredFeatures.HUGE_BLACK_MUSHROOM);
     }
@@ -29,6 +34,11 @@ public class BlackMushroomBlock extends MushroomBlock {
                     || Helper.isPosInVampireBiome(pos, (LevelAccessor) levelReader))
                     && blockState.canSustainPlant(levelReader, blockPos, net.minecraft.core.Direction.UP, this);
         }
+    }
+
+    @Override
+    public VoxelShape getShape(BlockState pState, BlockGetter pLevel, BlockPos pPos, CollisionContext pContext) {
+        return SHAPE;
     }
 
     @Override